created RecipeDetailView component
