+++ src/lib/RecipeCard/RecipeCard.svelte.test.ts
@@ -44,54 +44,6 @@ describe('RecipeCard', () => {
 		expect(screen.getByText(mockRecipe.description)).toBeInTheDocument();
 	});
 
-	it('renders nutrition information correctly', () => {
-		expect(screen.getByText('Nutrition')).toBeInTheDocument();
-
-		// Check calories (first value in nutrition array)
-		expect(screen.getByText('792')).toBeInTheDocument();
-
-		// Check PDV values
-		expect(screen.getByText('73% DV')).toBeInTheDocument(); // Total Fat
-		expect(screen.getByText('15% DV')).toBeInTheDocument(); // Sugar
-		expect(screen.getByText('46% DV')).toBeInTheDocument(); // Sodium
-		expect(screen.getByText('116% DV')).toBeInTheDocument(); // Protein
-		expect(screen.getByText('67% DV')).toBeInTheDocument(); // Saturated Fat
-
-		// Check labels
-		expect(screen.getByText('Calories:')).toBeInTheDocument();
-		expect(screen.getByText('Total Fat:')).toBeInTheDocument();
-		expect(screen.getByText('Sugar:')).toBeInTheDocument();
-		expect(screen.getByText('Sodium:')).toBeInTheDocument();
-		expect(screen.getByText('Protein:')).toBeInTheDocument();
-		expect(screen.getByText('Sat. Fat:')).toBeInTheDocument();
-	});
-
-	it('renders ingredients list correctly', () => {
-		expect(screen.getByText('Ingredients')).toBeInTheDocument();
-
-		// Check individual ingredients
-		expect(screen.getByText('boneless chicken')).toBeInTheDocument();
-		expect(screen.getByText('chicken flavor stuffing mix')).toBeInTheDocument();
-		expect(screen.getByText('cream of chicken soup')).toBeInTheDocument();
-		expect(screen.getByText('chicken broth')).toBeInTheDocument();
-	});
-
-	it('renders instructions correctly', () => {
-		expect(screen.getByText('Instructions')).toBeInTheDocument();
-
-		// Check that numbered steps are rendered
-		expect(screen.getByText('1')).toBeInTheDocument();
-		expect(screen.getByText('2')).toBeInTheDocument();
-		expect(screen.getByText('3')).toBeInTheDocument();
-		expect(screen.getByText('4')).toBeInTheDocument();
-
-		// Check step content
-		expect(screen.getByText(/empty can of chicken into large bowl/)).toBeInTheDocument();
-		expect(screen.getByText(/open stuffing and can of cream of chicken/)).toBeInTheDocument();
-		expect(screen.getByText(/place in crockpot/)).toBeInTheDocument();
-		expect(screen.getByText(/cook on medium for approx 2 hours/)).toBeInTheDocument();
-	});
-
 	it('has proper accessibility attributes', () => {
 		const card = screen.getByRole('article');
 		expect(card).toHaveAttribute('aria-label', `Recipe card for ${mockRecipe.name}`);
